Small Tableau rule change specifically aimed at Easthaven.
